Extends Export Support to MH Wilds
- Extends support for exporting MH Wilds textures
- Cleans up the texture code
- Fixes a heap corruption bug on the GDeflate wrapper
