+++ modules/tex/enums/dds_bpps.py
@@ -1,0 +1,128 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Fri Mar  7 16:26:14 2025
+
+@author: Asterisk
+"""
+
+ddsBpps = {
+	"UNKNOWN":0,
+	"R32G32B32A32TYPELESS":128,
+	"R32G32B32A32FLOAT":128,
+	"R32G32B32A32UINT":128,
+	"R32G32B32A32SINT":128,
+	"R32G32B32TYPELESS":96,
+	"R32G32B32FLOAT":96,
+	"R32G32B32UINT":96,
+	"R32G32B32SINT":96,
+	"R16G16B16A16TYPELESS":64,
+	"R16G16B16A16FLOAT":64,
+	"R16G16B16A16UNORM":64,
+	"R16G16B16A16UINT":64,
+	"R16G16B16A16SNORM":64,
+	"R16G16B16A16SINT":64,
+	"R32G32TYPELESS":64,
+	"R32G32FLOAT":64,
+	"R32G32UINT":64,
+	"R32G32SINT":64,
+	"R32G8X24TYPELESS":64,
+	"D32FLOATS8X24UINT":64,
+	"R32FLOATX8X24TYPELESS":64,
+	"X32TYPELESSG8X24UINT":64,
+	"R10G10B10A2TYPELESS":32,
+	"R10G10B10A2UNORM":32,
+	"R10G10B10A2UINT":32,
+	"R11G11B10FLOAT":32,
+	"R8G8B8A8TYPELESS":32,
+	"R8G8B8A8UNORM":32,
+	"R8G8B8A8UNORMSRGB":32,
+	"R8G8B8A8UINT":32,
+	"R8G8B8A8SNORM":32,
+	"R8G8B8A8SINT":32,
+	"R16G16TYPELESS":32,
+	"R16G16FLOAT":32,
+	"R16G16UNORM":32,
+	"R16G16UINT":32,
+	"R16G16SNORM":32,
+	"R16G16SINT":32,
+	"R32TYPELESS":32,
+	"D32FLOAT":32,
+	"R32FLOAT":32,
+	"R32UINT":32,
+	"R32SINT":32,
+	"R24G8TYPELESS":32,
+	"D24UNORMS8UINT":32,
+	"R24UNORMX8TYPELESS":32,
+	"X24TYPELESSG8UINT":32,
+	"R8G8TYPELESS":16,
+	"R8G8UNORM":16,
+	"R8G8UINT":16,
+	"R8G8SNORM":16,
+	"R8G8SINT":16,
+	"R16TYPELESS":16,
+	"R16FLOAT":16,
+	"D16UNORM":16,
+	"R16UNORM":16,
+	"R16UINT":16,
+	"R16SNORM":16,
+	"R16SINT":16,
+	"R8TYPELESS":8,
+	"R8UNORM":8,
+	"R8UINT":8,
+	"R8SNORM":8,
+	"R8SINT":8,
+	"A8UNORM":8,
+	"R1UNORM":1,
+	"R9G9B9E5SHAREDEXP":32,
+	"R8G8B8G8UNORM":16,
+	"G8R8G8B8UNORM":16,
+	"B5G6R5UNORM":16,
+	"B5G5R5A1UNORM":16,
+	"B8G8R8A8UNORM":32,
+	"B8G8R8X8UNORM":32,
+	"B8G8R8A8TYPELESS":32,
+	"R10G10B10XRBIASA2UNORM":32,
+	"B8G8R8A8UNORMSRGB":32,
+	"B8G8R8X8TYPELESS":32,
+	"B8G8R8X8UNORMSRGB":32,
+	"AYUV":32,
+	"Y410":10,
+	"Y416":16,
+	"NV12":12,
+	"P010":10,
+	"P016":16,
+	"DXGIFORMAT420OPAQUE":20,
+	"YUY2":16,
+	"Y210":10,
+	"Y216":16,
+	"NV11":11,
+	"AI44":44,
+	"IA44":44,
+	"P8":8,
+	"A8P8":16,
+	"B4G4R4A4UNORM":16,
+	"P208":8,
+	"V208":8,
+	"V408":8,
+	"BC1TYPELESS":4,
+	"BC1UNORM":4,
+	"BC1UNORMSRGB":4,
+	"BC2TYPELESS":8,
+	"BC2UNORM":8,
+	"BC2UNORMSRGB":8,
+	"BC3TYPELESS":8,
+	"BC3UNORM":8,
+	"BC3UNORMSRGB":8,
+	"BC4TYPELESS":4,
+	"BC4UNORM":4,
+	"BC4SNORM":4,
+	"BC5TYPELESS":8,
+	"BC5UNORM":8,
+	"BC5SNORM":8,
+	"BC6HTYPELESS":8,
+	"BC6HUF16":8,
+	"BC6HSF16":8,
+	"BC7TYPELESS":8,
+	"BC7UNORM":8,
+	"BC7UNORMSRGB":8,
+	}
