+++ modules/tex/enums/dxgi_format_enum.py
@@ -1,0 +1,244 @@
+# -*- coding: utf-8 -*-
+
+formatStringToDXGIDict = {
+	"UNKNOWN":0,
+	"R32G32B32A32TYPELESS":1,
+	"R32G32B32A32FLOAT":2,
+	"R32G32B32A32UINT":3,
+	"R32G32B32A32SINT":4,
+	"R32G32B32TYPELESS":5,
+	"R32G32B32FLOAT":6,
+	"R32G32B32UINT":7,
+	"R32G32B32SINT":8,
+	"R16G16B16A16TYPELESS":9,
+	"R16G16B16A16FLOAT":10,
+	"R16G16B16A16UNORM":11,
+	"R16G16B16A16UINT":12,
+	"R16G16B16A16SNORM":13,
+	"R16G16B16A16SINT":14,
+	"R32G32TYPELESS":15,
+	"R32G32FLOAT":16,
+	"R32G32UINT":17,
+	"R32G32SINT":18,
+	"R32G8X24TYPELESS":19,
+	"D32FLOATS8X24UINT":20,
+	"R32FLOATX8X24TYPELESS":21,
+	"X32TYPELESSG8X24UINT":22,
+	"R10G10B10A2TYPELESS":23,
+	"R10G10B10A2UNORM":24,
+	"R10G10B10A2UINT":25,
+	"R11G11B10FLOAT":26,
+	"R8G8B8A8TYPELESS":27,
+	"R8G8B8A8UNORM":28,
+	"R8G8B8A8UNORMSRGB":29,
+	"R8G8B8A8UINT":30,
+	"R8G8B8A8SNORM":31,
+	"R8G8B8A8SINT":32,
+	"R16G16TYPELESS":33,
+	"R16G16FLOAT":34,
+	"R16G16UNORM":35,
+	"R16G16UINT":36,
+	"R16G16SNORM":37,
+	"R16G16SINT":38,
+	"R32TYPELESS":39,
+	"D32FLOAT":40,
+	"R32FLOAT":41,
+	"R32UINT":42,
+	"R32SINT":43,
+	"R24G8TYPELESS":44,
+	"D24UNORMS8UINT":45,
+	"R24UNORMX8TYPELESS":46,
+	"X24TYPELESSG8UINT":47,
+	"R8G8TYPELESS":48,
+	"R8G8UNORM":49,
+	"R8G8UINT":50,
+	"R8G8SNORM":51,
+	"R8G8SINT":52,
+	"R16TYPELESS":53,
+	"R16FLOAT":54,
+	"D16UNORM":55,
+	"R16UNORM":56,
+	"R16UINT":57,
+	"R16SNORM":58,
+	"R16SINT":59,
+	"R8TYPELESS":60,
+	"R8UNORM":61,
+	"R8UINT":62,
+	"R8SNORM":63,
+	"R8SINT":64,
+	"A8UNORM":65,
+	"R1UNORM":66,
+	"R9G9B9E5SHAREDEXP":67,
+	"R8G8B8G8UNORM":68,
+	"G8R8G8B8UNORM":69,
+	"BC1TYPELESS":70,
+	"BC1UNORM":71,
+	"BC1UNORMSRGB":72,
+	"BC2TYPELESS":73,
+	"BC2UNORM":74,
+	"BC2UNORMSRGB":75,
+	"BC3TYPELESS":76,
+	"BC3UNORM":77,
+	"BC3UNORMSRGB":78,
+	"BC4TYPELESS":79,
+	"BC4UNORM":80,
+	"BC4SNORM":81,
+	"BC5TYPELESS":82,
+	"BC5UNORM":83,
+	"BC5SNORM":84,
+	"B5G6R5UNORM":85,
+	"B5G5R5A1UNORM":86,
+	"B8G8R8A8UNORM":87,
+	"B8G8R8X8UNORM":88,
+	"R10G10B10XRBIASA2UNORM":89,
+	"B8G8R8A8TYPELESS":90,
+	"B8G8R8A8UNORMSRGB":91,
+	"B8G8R8X8TYPELESS":92,
+	"B8G8R8X8UNORMSRGB":93,
+	"BC6HTYPELESS":94,
+	"BC6HUF16":95,
+	"BC6HSF16":96,
+	"BC7TYPELESS":97,
+	"BC7UNORM":98,
+	"BC7UNORMSRGB":99,
+	"AYUV":100,
+	"Y410":101,
+	"Y416":102,
+	"NV12":103,
+	"P010":104,
+	"P016":105,
+	"DXGIFORMAT420OPAQUE":106,
+	"YUY2":107,
+	"Y210":108,
+	"Y216":109,
+	"NV11":110,
+	"AI44":111,
+	"IA44":112,
+	"P8":113,
+	"A8P8":114,
+	"B4G4R4A4UNORM":115,
+	"P208":130,
+	"V208":131,
+	"V408":132,
+	"FORCEUINT":0xffffffff,
+	}
+DXGIToFormatStringDict = {
+	1: 'R32G32B32A32TYPELESS',
+	2: 'R32G32B32A32FLOAT',
+	3: 'R32G32B32A32UINT',
+	4: 'R32G32B32A32SINT',
+	5: 'R32G32B32TYPELESS',
+	6: 'R32G32B32FLOAT',
+	7: 'R32G32B32UINT',
+	8: 'R32G32B32SINT',
+	9: 'R16G16B16A16TYPELESS',
+	10: 'R16G16B16A16FLOAT',
+	11: 'R16G16B16A16UNORM',
+	12: 'R16G16B16A16UINT',
+	13: 'R16G16B16A16SNORM',
+	14: 'R16G16B16A16SINT',
+	15: 'R32G32TYPELESS',
+	16: 'R32G32FLOAT',
+	17: 'R32G32UINT',
+	18: 'R32G32SINT',
+	19: 'R32G8X24TYPELESS',
+	20: 'D32FLOATS8X24UINT',
+	21: 'R32FLOATX8X24TYPELESS',
+	22: 'X32TYPELESSG8X24UINT',
+	23: 'R10G10B10A2TYPELESS',
+	24: 'R10G10B10A2UNORM',
+	25: 'R10G10B10A2UINT',
+	26: 'R11G11B10FLOAT',
+	27: 'R8G8B8A8TYPELESS',
+	28: 'R8G8B8A8UNORM',
+	29: 'R8G8B8A8UNORMSRGB',
+	30: 'R8G8B8A8UINT',
+	31: 'R8G8B8A8SNORM',
+	32: 'R8G8B8A8SINT',
+	33: 'R16G16TYPELESS',
+	34: 'R16G16FLOAT',
+	35: 'R16G16UNORM',
+	36: 'R16G16UINT',
+	37: 'R16G16SNORM',
+	38: 'R16G16SINT',
+	39: 'R32TYPELESS',
+	40: 'D32FLOAT',
+	41: 'R32FLOAT',
+	42: 'R32UINT',
+	43: 'R32SINT',
+	44: 'R24G8TYPELESS',
+	45: 'D24UNORMS8UINT',
+	46: 'R24UNORMX8TYPELESS',
+	47: 'X24TYPELESSG8UINT',
+	48: 'R8G8TYPELESS',
+	49: 'R8G8UNORM',
+	50: 'R8G8UINT',
+	51: 'R8G8SNORM',
+	52: 'R8G8SINT',
+	53: 'R16TYPELESS',
+	54: 'R16FLOAT',
+	55: 'D16UNORM',
+	56: 'R16UNORM',
+	57: 'R16UINT',
+	58: 'R16SNORM',
+	59: 'R16SINT',
+	60: 'R8TYPELESS',
+	61: 'R8UNORM',
+	62: 'R8UINT',
+	63: 'R8SNORM',
+	64: 'R8SINT',
+	65: 'A8UNORM',
+	66: 'R1UNORM',
+	67: 'R9G9B9E5SHAREDEXP',
+	68: 'R8G8B8G8UNORM',
+	69: 'G8R8G8B8UNORM',
+	70: 'BC1TYPELESS',
+	71: 'BC1UNORM',
+	72: 'BC1UNORMSRGB',
+	73: 'BC2TYPELESS',
+	74: 'BC2UNORM',
+	75: 'BC2UNORMSRGB',
+	76: 'BC3TYPELESS',
+	77: 'BC3UNORM',
+	78: 'BC3UNORMSRGB',
+	79: 'BC4TYPELESS',
+	80: 'BC4UNORM',
+	81: 'BC4SNORM',
+	82: 'BC5TYPELESS',
+	83: 'BC5UNORM',
+	84: 'BC5SNORM',
+	85: 'B5G6R5UNORM',
+	86: 'B5G5R5A1UNORM',
+	87: 'B8G8R8A8UNORM',
+	88: 'B8G8R8X8UNORM',
+	89: 'R10G10B10XRBIASA2UNORM',
+	90: 'B8G8R8A8TYPELESS',
+	91: 'B8G8R8A8UNORMSRGB',
+	92: 'B8G8R8X8TYPELESS',
+	93: 'B8G8R8X8UNORMSRGB',
+	94: 'BC6HTYPELESS',
+	95: 'BC6HUF16',
+	96: 'BC6HSF16',
+	97: 'BC7TYPELESS',
+	98: 'BC7UNORM',
+	99: 'BC7UNORMSRGB',
+	100: 'AYUV',
+	101: 'Y410',
+	102: 'Y416',
+	103: 'NV12',
+	104: 'P010',
+	105: 'P016',
+	106: 'DXGIFORMAT420OPAQUE',
+	107: 'YUY2',
+	108: 'Y210',
+	109: 'Y216',
+	110: 'NV11',
+	111: 'AI44',
+	112: 'IA44',
+	113: 'P8',
+	114: 'A8P8',
+	115: 'B4G4R4A4UNORM',
+	130: 'P208',
+	131: 'V208',
+	132: 'V408',
+	4294967295: 'FORCEUINT'}
