+++ modules/tex/enums/game_version_enum.py
@@ -1,0 +1,20 @@
+# -*- coding: utf-8 -*-
+
+
+gameNameToTexVersionDict = {
+	"DMC5":11,
+	"RE2":10,
+	"RE3":190820018,
+	"MHR":28,
+	"MHRSB":28,
+	"RE8":30,
+	"RE2RT":34,
+	"RE3RT":34,
+	"RE7RT":35,
+	"RE4":143221013,
+	"SF6":143230113,
+	"DD2":760230703,
+	"KG":231106777,
+	"DR":240606151,
+	"MHWILDS":241106027,
+	}
