+++ modules/tex/enums/legacy_mappings.py
@@ -1,0 +1,36 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Fri Mar  7 17:34:42 2025
+
+@author: Asterisk
+"""
+
+#DDS To Tex
+DXT1 = 827611204  # 0x31545844
+DXT2 = 844388420  # 0x32545844
+DXT3 = 861165636  # 0x33545844
+DXT4 = 877942852  # 0x34545844
+DXT5 = 894720068  # 0x35545844
+ATI1 = 826889281
+ATI2 = 843666497
+BC4U = 1429488450
+BC4S = 1395934018
+BC5U = 1429553986
+BC5S = 1395999554
+DX10 = 808540228  # 0x30315844
+CCCC = 1128481603
+NULL = 0
+
+legacyMapping = {
+    DXT1: "BC1UNORM",
+    DXT2: "BC2UNORM",
+    DXT3: "BC2UNORMSRGB",
+    DXT4: "BC3UNORM",
+    DXT5: "BC3UNORMSRGB",
+    ATI1: "BC4UNORM",
+    ATI2: "BC5UNORM",
+    BC4U: "BC4UNORM",
+    BC4S: "BC4SNORM",
+    BC5U: "BC5UNORM",
+    BC5S: "BC5SNORM",
+}
