+++ modules/tex/enums/tex_format_enum.py
@@ -1,0 +1,266 @@
+# -*- coding: utf-8 -*-
+
+texFormatToDXGIStringDict = {
+		1:"R32G32B32A32TYPELESS",
+		2:"R32G32B32A32FLOAT",
+		3:"R32G32B32A32UINT",
+		4:"R32G32B32A32SINT",
+		5:"R32G32B32TYPELESS",
+		6:"R32G32B32FLOAT",
+		7:"R32G32B32UINT",
+		8:"R32G32B32SINT",
+		9:"R16G16B16A16TYPELESS",
+		0XA:"R16G16B16A16FLOAT",
+		0XB:"R16G16B16A16UNORM",
+		0XC:"R16G16B16A16UINT",
+		0XD:"R16G16B16A16SNORM",
+		0XE:"R16G16B16A16SINT",
+		0XF:"R32G32TYPELESS",
+		0X10:"R32G32FLOAT",
+		0X11:"R32G32UINT",
+		0X12:"R32G32SINT",
+		0X13:"R32G8X24TYPELESS",
+		0X14:"D32FLOATS8X24UINT",
+		0X15:"R32FLOATX8X24TYPELESS",
+		0X16:"X32TYPELESSG8X24UINT",
+		0X17:"R10G10B10A2TYPELESS",
+		0X18:"R10G10B10A2UNORM",
+		0X19:"R10G10B10A2UINT",
+		0X1A:"R11G11B10FLOAT",
+		0X1B:"R8G8B8A8TYPELESS",
+		0X1C:"R8G8B8A8UNORM",
+		0X1D:"R8G8B8A8UNORMSRGB",
+		0X1E:"R8G8B8A8UINT",
+		0X1F:"R8G8B8A8SNORM",
+		0X20:"R8G8B8A8SINT",
+		0X21:"R16G16TYPELESS",
+		0X22:"R16G16FLOAT",
+		0X23:"R16G16UNORM",
+		0X24:"R16G16UINT",
+		0X25:"R16G16SNORM",
+		0X26:"R16G16SINT",
+		0X27:"R32TYPELESS",
+		0X28:"D32FLOAT",
+		0X29:"R32FLOAT",
+		0X2A:"R32UINT",
+		0X2B:"R32SINT",
+		0X2C:"R24G8TYPELESS",
+		0X2D:"D24UNORMS8UINT",
+		0X2E:"R24UNORMX8TYPELESS",
+		0X2F:"X24TYPELESSG8UINT",
+		0X30:"R8G8TYPELESS",
+		0X31:"R8G8UNORM",
+		0X32:"R8G8UINT",
+		0X33:"R8G8SNORM",
+		0X34:"R8G8SINT",
+		0X35:"R16TYPELESS",
+		0X36:"R16FLOAT",
+		0X37:"D16UNORM",
+		0X38:"R16UNORM",
+		0X39:"R16UINT",
+		0X3A:"R16SNORM",
+		0X3B:"R16SINT",
+		0X3C:"R8TYPELESS",
+		0X3D:"R8UNORM",
+		0X3E:"R8UINT",
+		0X3F:"R8SNORM",
+		0X40:"R8SINT",
+		0X41:"A8UNORM",
+		0X42:"R1UNORM",
+		0X43:"R9G9B9E5SHAREDEXP",
+		0X44:"R8G8B8G8UNORM",
+		0X45:"G8R8G8B8UNORM",
+		0X46:"BC1TYPELESS",
+		0X47:"BC1UNORM",
+		0X48:"BC1UNORMSRGB",
+		0X49:"BC2TYPELESS",
+		0X4A:"BC2UNORM",
+		0X4B:"BC2UNORMSRGB",
+		0X4C:"BC3TYPELESS",
+		0X4D:"BC3UNORM",
+		0X4E:"BC3UNORMSRGB",
+		0X4F:"BC4TYPELESS",
+		0X50:"BC4UNORM",
+		0X51:"BC4SNORM",
+		0X52:"BC5TYPELESS",
+		0X53:"BC5UNORM",
+		0X54:"BC5SNORM",
+		0X55:"B5G6R5UNORM",
+		0X56:"B5G5R5A1UNORM",
+		0X57:"B8G8R8A8UNORM",
+		0X58:"B8G8R8X8UNORM",
+		0X59:"R10G10B10XRBIASA2UNORM",
+		0X5A:"B8G8R8A8TYPELESS",
+		0X5B:"B8G8R8A8UNORMSRGB",
+		0X5C:"B8G8R8X8TYPELESS",
+		0X5D:"B8G8R8X8UNORMSRGB",
+		0X5E:"BC6HTYPELESS",
+		0X5F:"BC6HUF16",
+		0X60:"BC6HSF16",
+		0X61:"BC7TYPELESS",
+		0X62:"BC7UNORM",
+		0X63:"BC7UNORMSRGB",
+		0X400:"VIAEXTENSION",
+		0X401:"ASTC4X4TYPELESS",
+		0X402:"ASTC4X4UNORM",
+		0X403:"ASTC4X4UNORMSRGB",
+		0X404:"ASTC5X4TYPELESS",
+		0X405:"ASTC5X4UNORM",
+		0X406:"ASTC5X4UNORMSRGB",
+		0X407:"ASTC5X5TYPELESS",
+		0X408:"ASTC5X5UNORM",
+		0X409:"ASTC5X5UNORMSRGB",
+		0X40A:"ASTC6X5TYPELESS",
+		0X40B:"ASTC6X5UNORM",
+		0X40C:"ASTC6X5UNORMSRGB",
+		0X40D:"ASTC6X6TYPELESS",
+		0X40E:"ASTC6X6UNORM",
+		0X40F:"ASTC6X6UNORMSRGB",
+		0X410:"ASTC8X5TYPELESS",
+		0X411:"ASTC8X5UNORM",
+		0X412:"ASTC8X5UNORMSRGB",
+		0X413:"ASTC8X6TYPELESS",
+		0X414:"ASTC8X6UNORM",
+		0X415:"ASTC8X6UNORMSRGB",
+		0X416:"ASTC8X8TYPELESS",
+		0X417:"ASTC8X8UNORM",
+		0X418:"ASTC8X8UNORMSRGB",
+		0X419:"ASTC10X5TYPELESS",
+		0X41A:"ASTC10X5UNORM",
+		0X41B:"ASTC10X5UNORMSRGB",
+		0X41C:"ASTC10X6TYPELESS",
+		0X41D:"ASTC10X6UNORM",
+		0X41E:"ASTC10X6UNORMSRGB",
+		0X41F:"ASTC10X8TYPELESS",
+		0X420:"ASTC10X8UNORM",
+		0X421:"ASTC10X8UNORMSRGB",
+		0X422:"ASTC10X10TYPELESS",
+		0X423:"ASTC10X10UNORM",
+		0X424:"ASTC10X10UNORMSRGB",
+		0X425:"ASTC12X10TYPELESS",
+		0X426:"ASTC12X10UNORM",
+		0X427:"ASTC12X10UNORMSRGB",
+		0X428:"ASTC12X12TYPELESS",
+		0X429:"ASTC12X12UNORM",
+		0X42A:"ASTC12X12UNORMSRGB",
+		0X7FFFFFFF:"FORCEUINT"
+		}
+formatStringToTexFormatDict = {
+	'R32G32B32A32TYPELESS': 1,
+	'R32G32B32A32FLOAT': 2,
+	'R32G32B32A32UINT': 3,
+	'R32G32B32A32SINT': 4,
+	'R32G32B32TYPELESS': 5,
+	'R32G32B32FLOAT': 6,
+	'R32G32B32UINT': 7,
+	'R32G32B32SINT': 8,
+	'R16G16B16A16TYPELESS': 9,
+	'R16G16B16A16FLOAT': 10,
+	'R16G16B16A16UNORM': 11,
+	'R16G16B16A16UINT': 12,
+	'R16G16B16A16SNORM': 13,
+	'R16G16B16A16SINT': 14,
+	'R32G32TYPELESS': 15,
+	'R32G32FLOAT': 16,
+	'R32G32UINT': 17,
+	'R32G32SINT': 18,
+	'R32G8X24TYPELESS': 19,
+	'D32FLOATS8X24UINT': 20,
+	'R32FLOATX8X24TYPELESS': 21,
+	'X32TYPELESSG8X24UINT': 22,
+	'R10G10B10A2TYPELESS': 23,
+	'R10G10B10A2UNORM': 24,
+	'R10G10B10A2UINT': 25,
+	'R11G11B10FLOAT': 26,
+	'R8G8B8A8TYPELESS': 27,
+	'R8G8B8A8UNORM': 28,
+	'R8G8B8A8UNORMSRGB': 29,
+	'R8G8B8A8UINT': 30,
+	'R8G8B8A8SNORM': 31,
+	'R8G8B8A8SINT': 32,
+	'R16G16TYPELESS': 33,
+	'R16G16FLOAT': 34,
+	'R16G16UNORM': 35,
+	'R16G16UINT': 36,
+	'R16G16SNORM': 37,
+	'R16G16SINT': 38,
+	'R32TYPELESS': 39,
+	'D32FLOAT': 40,
+	'R32FLOAT': 41,
+	'R32UINT': 42,
+	'R32SINT': 43,
+	'R24G8TYPELESS': 44,
+	'D24UNORMS8UINT': 45,
+	'R24UNORMX8TYPELESS': 46,
+	'X24TYPELESSG8UINT': 47,
+	'R8G8TYPELESS': 48,
+	'R8G8UNORM': 49,
+	'R8G8UINT': 50,
+	'R8G8SNORM': 51,
+	'R8G8SINT': 52,
+	'R16TYPELESS': 53,
+	'R16FLOAT': 54,
+	'D16UNORM': 55,
+	'R16UNORM': 56,
+	'R16UINT': 57,
+	'R16SNORM': 58,
+	'R16SINT': 59,
+	'R8TYPELESS': 60,
+	'R8UNORM': 61,
+	'R8UINT': 62,
+	'R8SNORM': 63,
+	'R8SINT': 64,
+	'A8UNORM': 65,
+	'R1UNORM': 66,
+	'R9G9B9E5SHAREDEXP': 67,
+	'R8G8B8G8UNORM': 68,
+	'G8R8G8B8UNORM': 69,
+	'BC1TYPELESS': 70,
+	'BC1UNORM': 71,
+	'BC1UNORMSRGB': 72,
+	'BC2TYPELESS': 73,
+	'BC2UNORM': 74,
+	'BC2UNORMSRGB': 75,
+	'BC3TYPELESS': 76,
+	'BC3UNORM': 77,
+	'BC3UNORMSRGB': 78,
+	'BC4TYPELESS': 79,
+	'BC4UNORM': 80,
+	'BC4SNORM': 81,
+	'BC5TYPELESS': 82,
+	'BC5UNORM': 83,
+	'BC5SNORM': 84,
+	'B5G6R5UNORM': 85,
+	'B5G5R5A1UNORM': 86,
+	'B8G8R8A8UNORM': 87,
+	'B8G8R8X8UNORM': 88,
+	'R10G10B10XRBIASA2UNORM': 89,
+	'B8G8R8A8TYPELESS': 90,
+	'B8G8R8A8UNORMSRGB': 91,
+	'B8G8R8X8TYPELESS': 92,
+	'B8G8R8X8UNORMSRGB': 93,
+	'BC6HTYPELESS': 94,
+	'BC6HUF16': 95,
+	'BC6HSF16': 96,
+	'BC7TYPELESS': 97,
+	'BC7UNORM': 98,
+	'BC7UNORMSRGB': 99,
+	'VIAEXTENSION': 1024,
+	'ASTC4X4TYPELESS': 1025,
+	'ASTC4X4UNORM': 1026,
+	'ASTC4X4UNORMSRGB': 1027,
+	'ASTC5X4TYPELESS': 1028,
+	'ASTC5X4UNORM': 1029,
+	'ASTC5X4UNORMSRGB': 1030,
+	'ASTC5X5TYPELESS': 1031,
+	'ASTC5X5UNORM': 1032,
+	'ASTC5X5UNORMSRGB': 1033,
+	'ASTC6X5TYPELESS': 1034,
+	'ASTC6X5UNORM': 1035,
+	'ASTC6X5UNORMSRGB': 1036,
+	'ASTC6X6TYPELESS': 1037,
+	'ASTC6X6UNORM': 1038,
+	'ASTC6X6UNORMSRGB': 1039,
+	'ASTC8X5TYPELESS': 1040,
+	'ASTC8X5UNORM': 1041,
+}
