+++ re-mesh-editor_updater/RE-Mesh-Editor_updater_status.json
@@ -1,0 +1,9 @@
+{
+    "last_check": "",
+    "backup_date": "",
+    "update_ready": false,
+    "ignore": false,
+    "just_restored": false,
+    "just_updated": false,
+    "version_text": {}
+}
